weatherMap api fetch func moved to utils/API/openWeatherMap.js
