+++ client/src/DarkMode.jsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 // import IconButton from "@mui/material/IconButton";
-// import Box from "@mui/material/Box";
 import { ThemeProvider, createTheme } from "@mui/material/styles";
 // import Brightness4Icon from "@mui/icons-material/Brightness4";
 // import Brightness7Icon from "@mui/icons-material/Brightness7";
@@ -16,18 +15,7 @@ import { useMediaQuery } from "@mui/material";
 //   const theme = useTheme();
 //   const colorMode = React.useContext(ColorModeContext);
 //   return (
-//     <Box
-//       sx={{
-//         display: "flex",
-//         width: "100%",
-//         alignItems: "center",
-//         justifyContent: "center",
-//         bgcolor: "background.default",
-//         color: "text.primary",
-//         borderRadius: 1,
-//         p: 3,
-//       }}
-//     >
+
 //       {theme.palette.mode} mode
 //       <IconButton
 //         sx={{ ml: 1 }}
@@ -40,7 +28,6 @@ import { useMediaQuery } from "@mui/material";
 //           <Brightness4Icon />
 //         )}
 //       </IconButton>
-//     </Box>
 //   );
 // }
